Use `$return_status` to avoid zsh conflict
Prior to this change the test code used a `$status` variable to track
the return code from the `run` helper function. Unfortunately, `$status`
is used internally in ZSH as an alias for `$?` and reports to the user
as a read-only variable.

See: http://zsh.sourceforge.net/Doc/Release/Parameters.html#index-status

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>