Support ksh93 with its strict `local` support
ksh93 will fail when the `local` keyword is used and its function
variables are essentially global. This change shims in a `local`
function that will minimally satisfy ksh93 without affecting the other
shell interpreters. The use of `local` is very precise in this change to
not use the initialization form (i.e. `local <var>=<val>`) which allows
the other shell interpreters to use `local` if supported, ignore it if
it does not use the keyword directly, and shim it for ksh93. Note that
since ksh93 will leak its variables outside the function, the local
variables were renamed to use a leading underscore (to denote an
internal variable), and they are `unset` at the end of each function.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>