[install.sh] Vendor libsh.sh with reasonable perms
Prior to this change the temporary copy of libsh.sh was copied into the
destination, but because the temporary copy with created via `mktemp`,
the permissions were very restricted. This change sets permissions
indirectly using `touch` to create the target file if it isn't present,
relying on the current `umask` value when creating an empty file.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>