[install] Remove errant `s` from `set` if it exists
This situation was discovered when running the `install.sh` program via
a "curl/sh" invocation such as:

```console
$ curl --proto '=https' --tlsv1.2 -sSf \
    https://raw.githubusercontent.com/fnichol/libsh/master/install.sh \
  | sh -s -- --mode=vendor
```

In this case, an `s` was inserted in `$-` causing an error such as `set:
-s: invalid option` when calling the `download` function.

This change strips out any `s` characters before restoring the original
set flags.

Note that it is hypothesized that using `sh -s --` is causing this
behavior (directly invoking the script or not providing any arguments to
the piped `sh` work fine). It is not conclusive and therefore a little
troubling to have found this issue.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>