Update `die` to exit & introduce `check_cmd`
This changeset was spurred by a realization that a shell function which
returns a non-zero exit code will **not** exit the program or script
currently running, despite whether or not `set -e` is enabled. As `die`
in its name implies that the process should abort/terminate, an `exit 1`
is not used to replace the prior behavior of `return 1`.

So, `die` will no longer return from its function but rather terminate
the running shell process. Note that this should be a script but if
`die` is invoked in an interactive shell session, then the shell will
quit.

Given this update, the behavior of `need_cmd` also changed as a result
of using `die` under the hood. As a consequence, `need_cmd` will also no
longer return from its function if the given program could not be found
on `PATH`.

Finally as a result of all of the above, it seemed useful to have
another function that would only *check* for a program but not terminate
the script. Hence, `check_cmd` was introduced to return either `0` on
success or `1` on failure. The implementation of `need_cmd` was updated
to call `check_cmd` and `die` is still used for `need_cmd`.

Note that this is a **breaking change** for the following functions:

* `die`: will quit the running shell and not return
* `need_cmd`: will quit the running shell and not return if the program
  cannot be found

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>