Add `setup_traps` function to libsh
The `setup_traps` function will set up a trap for `EXIT` as well as
common signals such as `HUP`, `INT`, `QUIT`, `ALRM`, and `TERM`. This
implementation will use the same "cleanup function" for all traps which
may or may not be the desired behavior depending on the script's
implementation needs.

Note that the Public Domain Korn Shell (i.e. pdksh) has 3 pending signal
tests that are skipped. While it appears that the implementation works
and the method of signaling a backgrounded script works, it has been a
challenge to get the unit tests to pass.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>