[install] Split a `local` declarations into their own lines
This change is an attempt to maximize the portability of `install.sh`
across as many shell implementations as libsh supports, namely
including `ksh` which technically does not parse `local`. libsh supports
strict local declarations via a shim provided when running under Ksh and
this change aims to make that supportable.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>