refactor: remove locals in functions taking 1 arg
This change removes local variables in functions that only use one
argument. While in theory this should lead to easier to read code, in
practice there is more safe setup and tearing down of these variables
than there is implementation logic in the function.

Instead where it makes sense, we'll force the reader to trace through
and follow one variable, `$1`, through the function body.

The net effect is that the line count of source code is decreased which
should help when shipping versions of this library in "curl-bash" style
scripts.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>