fix: support multiple version bumping scenarios
The default case assumes a `-dev` pre-release is set in `VERSION.txt` and
will simply remove the pre-release section of the version string. So if
the current version is `0.5.0-dev` then the default version bump will
result in `0.5.0`.

Alternatively, if a more drastic version bump is required, such as when
bumping the major version, then set `BUMP_MODE` to either `patch`,
`minor`, or `major`. For example a current version of `0.5.0-dev` with
`BUMP_MODE=major` would yield `1.0.0`.

If the desired version is a pre-release, such as an alpha, beta, or
release candidate, then set `BUMP_MODE=set` and set `BUMP_PRE` to to the
pre-release part. For example, a current version of `0.5.0` with
`BUMP_MODE=set` and `BUMP_PRE=rc.1` would yield `0.5.0-rc.1`.

Finally, if a specific version must be set, simply set `BUMP_SET`.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>