fix: ensure file/directory cleanups only apply to current process
This change fixes an issue where a sub-process or calling script which
also uses libsh will potentially clean up all files and/or directories
of the parent when the child's trap handlers fire. This is not expected
behavior--each process should essentially have its own record of files
and directories for clean up.

In order to ensure this happens, 2 new global variables are used and
tracked in `setup_cleanup_files` and `setup_cleanup_directories`:
`__CLEANUP_FILES_SETUP__` and `__CLEANUP_DIRECTORIES_SETUP__`
respectively.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>